+++ example/lib/main.dart
@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:flutter_multi_chip_select/flutter_multi_chip_select.dart';
 
 void main() => runApp(MyApp());
@@ -12,33 +9,9 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
-
-  @override
-  void initState() {
-    super.initState();
-    initPlatformState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    try {
-      platformVersion = await FlutterMultiChipSelect.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
-  }
+  final multiSelectKey = GlobalKey<MultiSelectDropdownState>();
+  String _result = "";
+  var menuItems = [1, 2, 3, 4, 5, 6];
 
   @override
   Widget build(BuildContext context) {
@@ -47,8 +20,49 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+        body: Column(
+          children: <Widget>[
+            Center(
+              child: Row(
+                children: <Widget>[
+                  Expanded(
+                    child: FlutterMultiChipSelect(
+                      key: multiSelectKey,
+                      elements: List.generate(
+                        menuItems.length,
+                        (index) => MultiSelectItem<String>.simple(
+                            actions: [
+                              IconButton(
+                                icon: Icon(Icons.delete),
+                                onPressed: () {
+                                  setState(() {
+                                    menuItems.remove(menuItems[index]);
+                                  });
+                                  print("Delete Call at: " + menuItems[index].toString());
+                                },
+                              )
+                            ],
+                            title: "Item " + menuItems[index].toString(),
+                            value: menuItems[index].toString()),
+                      ),
+                      label: "Dropdown Select",
+                      values: [],
+                    ),
+                  ),
+                  IconButton(
+                    icon: Icon(Icons.save),
+                    onPressed: () {
+                      setState(() {
+                        _result =
+                            this.multiSelectKey.currentState.result.toString();
+                      });
+                    },
+                  ),
+                ],
+              ),
+            ),
+            Text((_result != "") ? "Save Result" + _result : "")
+          ],
         ),
       ),
     );